refactor: 컴포넌트 타입을 스타일 컴포넌트에서 부르는 순환 참조 문제 해결
