+++ components/src/lib/components/Modal/Modal.tsx
@@ -1,11 +1,6 @@
-import { ReactNode } from "react";
 import { StyledModal } from "./Modal.styled";
 
-export interface ModalProps {
-  isOpen: boolean;
-  children: ReactNode;
-  position?: "bottom" | "center";
-}
+import { ModalProps } from "./Modal.types";
 
 const Modal = ({ children, ...props }: ModalProps) => {
   return <StyledModal {...props}>{children}</StyledModal>;
