+++ components/src/lib/components/Modal/Modal.types.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from "react";
+
+export interface ModalProps {
+  isOpen: boolean;
+  children: ReactNode;
+  position?: "bottom" | "center";
+}
