+++ components/src/lib/components/ModalContent/ModalContent.tsx
@@ -1,9 +1,6 @@
 import { StyledModalContent } from "./ModalContent.styled";
 
-export interface ModalContentProps {
-  children: React.ReactNode;
-  position?: "center" | "bottom";
-}
+import { ModalContentProps } from "./ModalContent.types";
 
 const ModalContent = ({ children, position = "center" }: ModalContentProps) => {
   return (
