+++ components/src/lib/components/ModalContent/ModalContent.types.ts
@@ -1,0 +1,4 @@
+export interface ModalContentProps {
+  children: React.ReactNode;
+  position?: "center" | "bottom";
+}
