+++ components/src/lib/components/ModalFooter/ModalFooter.tsx
@@ -1,13 +1,6 @@
-import { ReactNode } from "react";
-
 import { StyledModalFooter } from "./ModalFooter.styled";
 
-export interface ModalFooterProps {
-  children: ReactNode;
-  direction?: "row" | "column";
-  align?: "start" | "center" | "end";
-  justify?: "start" | "center" | "end";
-}
+import { ModalFooterProps } from "./ModalFooter.types";
 
 const ModalFooter = ({ children, ...props }: ModalFooterProps) => {
   return <StyledModalFooter {...props}>{children}</StyledModalFooter>;
