+++ components/src/lib/components/ModalFooter/ModalFooter.types.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+
+export interface ModalFooterProps {
+  children: ReactNode;
+  direction?: "row" | "column";
+  align?: "start" | "center" | "end";
+  justify?: "start" | "center" | "end";
+}
