+++ components/src/lib/components/ModalHeader/ModalHeader.tsx
@@ -1,12 +1,6 @@
-import { ReactNode } from "react";
 import { StyledModalHeader } from "./ModalHeader.styled";
 
-export interface ModalHeaderProps {
-  children: ReactNode;
-  direction?: "row" | "column";
-  align?: "start" | "center" | "end";
-  justify?: "start" | "center" | "end";
-}
+import { ModalHeaderProps } from "./ModalHeader.types";
 
 const ModalHeader = ({ children, ...props }: ModalHeaderProps) => {
   return <StyledModalHeader {...props}>{children}</StyledModalHeader>;
