+++ components/src/lib/components/ModalHeader/ModalHeader.types.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+
+export interface ModalHeaderProps {
+  children: ReactNode;
+  direction?: "row" | "column";
+  align?: "start" | "center" | "end";
+  justify?: "start" | "center" | "end";
+}
