+++ components/src/lib/components/ModalTitle/ModalTitle.types.ts
@@ -1,0 +1,8 @@
+import { JSX, ReactNode } from "react";
+
+export interface ModalTitleProps {
+  children: ReactNode;
+  fontSize: string;
+  fontWeight: string;
+  tag?: keyof JSX.IntrinsicElements;
+}
